Can now scroll through trials
Modified the load button function to also plot the first trial on that
day and set up the trial slider parameters to have the correct max
value/step size.
Also gave the slider buttonpress the ability to plot the correct trial
(according to slider position).
Gave the trial number edit textbox the ability to trigger plotting the
correct trial, or to tell the user that the requested trial does not
exist.